requirements/apps.txt
Django==1.2.4
South
python-dateutil
django-extensions
MySQL-python
markdown
django_generic_flatblocks
simplejson

